LPS-192496 Extracted service ConfigurationModelIndexer and move methods for ClusterConfigurationModelIndexer
