+++ modules/apps/configuration-admin/configuration-admin-web/src/main/java/com/liferay/configuration/admin/web/internal/search/ConfigurationModelIndexer.java
@@ -45,28 +45,23 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.ResourceBundle;
-import java.util.concurrent.ConcurrentHashMap;
 
 import javax.portlet.PortletRequest;
 import javax.portlet.PortletResponse;
 
-import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
-import org.osgi.framework.BundleEvent;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 import org.osgi.service.metatype.AttributeDefinition;
 import org.osgi.service.metatype.ObjectClassDefinition;
-import org.osgi.util.tracker.BundleTracker;
-import org.osgi.util.tracker.BundleTrackerCustomizer;
 
 /**
  * @author Michael C. Han
  */
 @Component(
 	property = {"index.on.startup=false", "system.index=true"},
-	service = {ConfigurationModelIndexer.class, Indexer.class}
+	service = Indexer.class
 )
 public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 
@@ -98,44 +93,6 @@ public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 		catch (Exception exception) {
 			throw new SearchException(exception);
 		}
-	}
-
-	public BundleTracker<Collection<ConfigurationModel>> initialize() {
-		Map<String, Collection<ConfigurationModel>> configurationModelsMap =
-			new ConcurrentHashMap<>();
-
-		Bundle[] bundles = _bundleContext.getBundles();
-
-		List<ConfigurationModel> configurationModelsList = new ArrayList<>();
-
-		for (Bundle bundle : bundles) {
-			if (bundle.getState() != Bundle.ACTIVE) {
-				continue;
-			}
-
-			Map<String, ConfigurationModel> configurationModels =
-				_configurationModelRetriever.getConfigurationModels(
-					bundle, ExtendedObjectClassDefinition.Scope.SYSTEM, null);
-
-			configurationModelsList.addAll(configurationModels.values());
-
-			configurationModelsMap.put(
-				bundle.getSymbolicName(), configurationModels.values());
-		}
-
-		reindex(configurationModelsList);
-
-		_commit();
-
-		BundleTracker<Collection<ConfigurationModel>> bundleTracker =
-			new BundleTracker<>(
-				_bundleContext, Bundle.ACTIVE,
-				new ConfigurationModelsBundleTrackerCustomizer(
-					configurationModelsMap));
-
-		bundleTracker.open();
-
-		return bundleTracker;
 	}
 
 	@Override
@@ -392,17 +349,6 @@ public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 		}
 	}
 
-	private void _commit() {
-		try {
-			_indexWriterHelper.commit();
-		}
-		catch (SearchException searchException) {
-			if (_log.isWarnEnabled()) {
-				_log.warn("Unable to commit", searchException);
-			}
-		}
-	}
-
 	private List<String> _getLocalizedValues(
 		List<String> attributeDescriptions,
 		ResourceBundleLoader resourceBundleLoader, Locale locale) {
@@ -502,76 +448,6 @@ public class ConfigurationModelIndexer extends BaseIndexer<ConfigurationModel> {
 		private final String _key;
 		private final String _name;
 		private final Map<Locale, String> _values = new HashMap<>();
-
-	}
-
-	private class ConfigurationModelsBundleTrackerCustomizer
-		implements BundleTrackerCustomizer<Collection<ConfigurationModel>> {
-
-		@Override
-		public Collection<ConfigurationModel> addingBundle(
-			Bundle bundle, BundleEvent bundleEvent) {
-
-			Collection<ConfigurationModel> configurationModels =
-				_configurationModelsMap.remove(bundle.getSymbolicName());
-
-			if (configurationModels != null) {
-				if (configurationModels.isEmpty()) {
-					return null;
-				}
-
-				return configurationModels;
-			}
-
-			Map<String, ConfigurationModel> configurationModelsMap =
-				_configurationModelRetriever.getConfigurationModels(
-					bundle, ExtendedObjectClassDefinition.Scope.SYSTEM, null);
-
-			if (configurationModelsMap.isEmpty()) {
-				return null;
-			}
-
-			reindex(configurationModelsMap.values());
-
-			_commit();
-
-			return configurationModelsMap.values();
-		}
-
-		@Override
-		public void modifiedBundle(
-			Bundle bundle, BundleEvent bundleEvent,
-			Collection<ConfigurationModel> configurationModels) {
-		}
-
-		@Override
-		public void removedBundle(
-			Bundle bundle, BundleEvent bundleEvent,
-			Collection<ConfigurationModel> configurationModels) {
-
-			for (ConfigurationModel configurationModel : configurationModels) {
-				try {
-					delete(configurationModel);
-				}
-				catch (SearchException searchException) {
-					if (_log.isWarnEnabled()) {
-						_log.warn("Unable to reindex models", searchException);
-					}
-				}
-			}
-
-			_commit();
-		}
-
-		private ConfigurationModelsBundleTrackerCustomizer(
-			Map<String, Collection<ConfigurationModel>>
-				configurationModelsMap) {
-
-			_configurationModelsMap = configurationModelsMap;
-		}
-
-		private final Map<String, Collection<ConfigurationModel>>
-			_configurationModelsMap;
 
 	}
